requirements.txt
Django>=5.0
djangorestframework
psycopg2-binary
django-environ
djangorestframework-simplejwt
drf-spectacular
drf-spectacular-sidecar
django-allauth
dj-rest-auth
requests
attrs
certifi
cffi
charset-normalizer
cryptography
idna
inflection
oauthlib
PyJWT
PyYAML
sqlparse
typing_extensions
urllib3

